+++ config/toml.go
@@ -73,8 +73,8 @@ gui_listen_addr = "{{ .BaseConfig.GUIListenAddress }}"
 # Address to listen for API connections
 api_listen_addr = "{{ .BaseConfig.APIListenAddress }}"
 
-# Enable events for API. Slows down node.
-enable_events = {{ .BaseConfig.EnableEvents }}
+# Sets node to be in validator mode. Disables API, events, history of blocks, indexes, etc. 
+validator_mode = {{ .BaseConfig.ValidatorMode }}
 
 # If this node is many blocks behind the tip of the chain, FastSync
 # allows them to catchup quickly by downloading blocks in parallel
@@ -200,28 +200,6 @@ size = {{ .Mempool.Size }}
 
 # size of the cache (used to filter transactions we saw earlier)
 cache_size = {{ .Mempool.CacheSize }}
-
-##### transactions indexer configuration options #####
-[tx_index]
-
-# What indexer to use for transactions
-#
-# Options:
-#   1) "null" (default)
-#   2) "kv" - the simplest possible indexer, backed by key-value storage (defaults to levelDB; see DBBackend).
-indexer = "{{ .TxIndex.Indexer }}"
-
-# Comma-separated list of tags to index (by default the only tag is tx hash)
-#
-# It's recommended to index only a subset of tags due to possible memory
-# bloat. This is, of course, depends on the indexer's DB and the volume of
-# transactions.
-index_tags = "{{ .TxIndex.IndexTags }}"
-
-# When set to true, tells indexer to index all tags. Note this may be not
-# desirable (see the comment above). IndexTags has a precedence over
-# IndexAllTags (i.e. when given both, IndexTags will be indexed).
-index_all_tags = {{ .TxIndex.IndexAllTags }}
 
 ##### instrumentation configuration options #####
 [instrumentation]