+++ core/state/state_object.go
@@ -17,41 +17,13 @@
 package state
 
 import (
-	"bytes"
-	"fmt"
 	"io"
 	"math/big"
 
 	"github.com/MinterTeam/minter-go-node/core/types"
-	"github.com/MinterTeam/minter-go-node/crypto"
 	"github.com/MinterTeam/minter-go-node/rlp"
 	"sort"
 )
-
-type Code []byte
-
-func (self Code) String() string {
-	return string(self) //strings.Join(Disassemble(self), " ")
-}
-
-type Storage map[types.Hash]types.Hash
-
-func (self Storage) String() (str string) {
-	for key, value := range self {
-		str += fmt.Sprintf("%X : %X\n", key, value)
-	}
-
-	return
-}
-
-func (self Storage) Copy() Storage {
-	cpy := make(Storage)
-	for key, value := range self {
-		cpy[key] = value
-	}
-
-	return cpy
-}
 
 // stateObject represents an Ethereum account which is being modified.
 //
@@ -60,23 +32,8 @@ func (self Storage) Copy() Storage {
 // Account values can be accessed and modified through the object.
 // Finally, call CommitTrie to write the modified storage trie into a database.
 type stateObject struct {
-	address  types.Address
-	addrHash types.Hash // hash of ethereum address of the account
-	data     Account
-	db       *StateDB
-
-	// DB error.
-	// State objects are used by the consensus core and VM which are
-	// unable to deal with database-level errors. Any error that occurs
-	// during a database read is memoized here and will eventually be returned
-	// by StateDB.Commit.
-	dbErr error
-
-	// Write caches.
-	trie Trie // storage trie, which becomes non-nil on first access
-
-	cachedStorage Storage // Storage entry cache to avoid duplicate reads
-	dirtyStorage  Storage // Storage entries that need to be flushed to disk
+	address types.Address
+	data    Account
 
 	// Cache flags.
 	// When an object is marked suicided it will be delete from the trie
@@ -162,26 +119,15 @@ func newObject(db *StateDB, address types.Address, data Account, onDirty func(ad
 	}
 
 	return &stateObject{
-		db:            db,
-		address:       address,
-		addrHash:      crypto.Keccak256Hash(address[:]),
-		data:          data,
-		cachedStorage: make(Storage),
-		dirtyStorage:  make(Storage),
-		onDirty:       onDirty,
+		address: address,
+		data:    data,
+		onDirty: onDirty,
 	}
 }
 
 // EncodeRLP implements rlp.Encoder.
 func (c *stateObject) EncodeRLP(w io.Writer) error {
 	return rlp.Encode(w, c.data)
-}
-
-// setError remembers the first non-nil error it is called with.
-func (self *stateObject) setError(err error) {
-	if self.dbErr == nil {
-		self.dbErr = err
-	}
 }
 
 func (self *stateObject) markSuicided() {
@@ -198,94 +144,6 @@ func (c *stateObject) touch() {
 		c.onDirty = nil
 	}
 	c.touched = true
-}
-
-func (c *stateObject) getTrie(db Database) Trie {
-	if c.trie == nil {
-		var err error
-		c.trie, err = db.OpenStorageTrie(c.addrHash, c.data.Root)
-		if err != nil {
-			c.trie, _ = db.OpenStorageTrie(c.addrHash, types.Hash{})
-			c.setError(fmt.Errorf("can't create storage trie: %v", err))
-		}
-	}
-	return c.trie
-}
-
-// GetState returns a value in account storage.
-func (self *stateObject) GetState(db Database, key types.Hash) types.Hash {
-	value, exists := self.cachedStorage[key]
-	if exists {
-		return value
-	}
-	// Load from DB in case it is missing.
-	enc, err := self.getTrie(db).TryGet(key[:])
-	if err != nil {
-		self.setError(err)
-		return types.Hash{}
-	}
-	if len(enc) > 0 {
-		_, content, _, err := rlp.Split(enc)
-		if err != nil {
-			self.setError(err)
-		}
-		value.SetBytes(content)
-	}
-	if (value != types.Hash{}) {
-		self.cachedStorage[key] = value
-	}
-	return value
-}
-
-// SetState updates a value in account storage.
-func (self *stateObject) SetState(db Database, key, value types.Hash) {
-	self.setState(key, value)
-}
-
-func (self *stateObject) setState(key, value types.Hash) {
-	self.cachedStorage[key] = value
-	self.dirtyStorage[key] = value
-
-	if self.onDirty != nil {
-		self.onDirty(self.Address())
-		self.onDirty = nil
-	}
-}
-
-// updateTrie writes cached storage modifications into the object's storage trie.
-func (self *stateObject) updateTrie(db Database) Trie {
-	tr := self.getTrie(db)
-	for key, value := range self.dirtyStorage {
-		delete(self.dirtyStorage, key)
-		if (value == types.Hash{}) {
-			self.setError(tr.TryDelete(key[:]))
-			continue
-		}
-		// Encoding []byte cannot fail, ok to ignore the error.
-		v, _ := rlp.EncodeToBytes(bytes.TrimLeft(value[:], "\x00"))
-		self.setError(tr.TryUpdate(key[:], v))
-	}
-	return tr
-}
-
-// UpdateRoot sets the trie root to the current root hash of
-func (self *stateObject) updateRoot(db Database) {
-	self.updateTrie(db)
-	self.data.Root = self.trie.Hash()
-}
-
-// CommitTrie the storage trie of the object to dwb.
-// This updates the trie root.
-func (self *stateObject) CommitTrie(db Database) error {
-	self.updateTrie(db)
-	if self.dbErr != nil {
-		return self.dbErr
-	}
-	root, err := self.trie.Commit(nil)
-	if err == nil {
-		self.data.Root = root
-	}
-	return err
 }
 
 // AddBalance removes amount from c's balance.
@@ -335,11 +193,6 @@ func (c *stateObject) ReturnGas(gas *big.Int) {}
 
 func (self *stateObject) deepCopy(db *StateDB, onDirty func(addr types.Address)) *stateObject {
 	stateObject := newObject(db, self.address, self.data, onDirty)
-	if self.trie != nil {
-		stateObject.trie = db.db.CopyTrie(self.trie)
-	}
-	stateObject.dirtyStorage = self.dirtyStorage.Copy()
-	stateObject.cachedStorage = self.dirtyStorage.Copy()
 	stateObject.suicided = self.suicided
 	stateObject.deleted = self.deleted
 	return stateObject